refactor: MenuGroup package 분리
