+++ src/main/java/kitchenpos/menu_group/application/MenuGroupService.java
@@ -1,9 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.menu_group.application;
 
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.dto.MenuGroupRequestDto;
-import kitchenpos.dto.MenuGroupResponseDto;
-import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.domain.MenuGroupRepository;
+import kitchenpos.menu_group.dto.MenuGroupRequestDto;
+import kitchenpos.menu_group.dto.MenuGroupResponseDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
