+++ src/main/java/kitchenpos/menu_group/domain/MenuGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.repository;
+package kitchenpos.menu_group.domain;
 
-import kitchenpos.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
