+++ src/main/java/kitchenpos/menu_group/dto/MenuGroupRequestDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.menu_group.dto;
 
 public class MenuGroupRequestDto {
