+++ src/test/java/kitchenpos/menu_group/domain/MenuGroupTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu_group.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
